refactored schema out to the top level
